The Idea
To resolve the above stated problem we have come with the Idea of Rent-o-Flix : an OTT platform which is an integration of all major OTT platforms like
Netflix, Amazon Prime Video, Hotstar, Hulu etc. so, the user should not toggle between apps, so users to avoid the confusion which to subscribe and pay
Secondly , we run algorithm, videos and series will be highlighted each platform according to the user choice. Moreover user can avail services on prepaid basis like hourly, daily, monthly, quarterly, and yearly as per user choice similar like mobile prepaid